frontend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
dj-database-url==2.1.0
Django==4.2.5
django-appconf==1.0.5
django-cockroachdb==4.2
django-cors-headers==4.3.1
idna==3.4
mailjet-rest==1.3.4
numpy==1.26.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
resend==0.7.2
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
